Refactor components into separate folder!
